Initial draft for testing overhaul.
An initial attempt to overhaul the `std.testing` API for writing
assertions in tests.

This is a first draft with the goal of reimplementing
`std.testing.expect` to return assertion objects for the given actual
value, according to the operations that can be performed on the type of
the actual value, effectively making `expect` a fluent assertions API.

Backward compatibility with the old version, `fn expect(ok: bool) void`
is a goal of this overhaul to ensure that not code is broken by the
change.

This overhaul aims to solve 2 problems with the existing implementation
of `expectEqual` and `expectEqualSlices`.

First, is the ambiguity in the ordering of arguments.  It is easy to
have the wrong ordering for the actual and expected value.

Second, the signature of `expectEqual(actual: var, expected:
@TypeOf(actual)) void` is problematic when the type of the expected is
an integer or float, as the actual must always be explicitly cast to the
actual type:

```zig
fn getAnswer(): i32 { return 42; }

test "everything" {
  // Here, the @as cast is required, because otherwise, actual will have
  // a type of `comptime_int`.
  std.testing.expectEqual(@as(i32, 42), getAnswer());
}
```