Fix generated docs of user code
fff3c1fff4c3ebfcb2bd4f08a43ae7815b5c446b broke the docs generation for
own user code. This adds a workaround, because all the code in docs/main.js
is relying on the fact that the root package of the main package is the
root package itself.