Add std.fs.path.resolveat fn and use it in CacheHash
This commit adds `std.fs.path.resolveat` functions which is similar
to `std.fs.path.resolve`, however, differs in the sense that, in
addition to a slice of paths, it also accepts a `Dir` handle with
respect to which the paths should be resolved. It also issues
one syscall in order to workout the path to the `Dir` handle.

With `std.fs.path.resolveat`, it is now possible to store `work_dir`
handle directly in `CacheHash` struct and workout paths to cached
files wrt to it rather than wrt `cwd`.