Add std.fs.path.resolveRelative helper fn
`std.fs.path.resolveRelative` function is like `std.fs.path.resolve`
however it resolves relative paths **only**, and they are always
resolved to wrt ".". If the final path is resolved to "." or beyond,
`null` is returned instead to signify this. Also, if an absolute
path is passed as an argument, `ResolveRelativeError.AbsolutePath`
is thrown. This function is particularly useful for Capsicum-like
targets such as WASI.

This commit also tweaks `std.cache_hash.CacheHash` to use either
`std.fs.path.resolveat` or `std.fs.path.resolveRelative` (the latter
used when targeting WASI). This way, we essentially got rid of any
mention of CWD, and hence, it is now possible to use `CacheHash`
struct in WASI. As a result, all tests in the said module have
been enabled for WASI.