Add std.fs.File.readToEnd convenience function
This commit adds a convenience function for reading the entire
contents of the file into a runtime-allocated buffer, `std.fs.File.readToEnd`.
This function is similar to Rust's `read_to_end` function.

Implementation-wise, it wraps `File.getEndPos` for extracting the file
size and `File.readAllAlloc` with `max_bytes == file_size`. If the
`file_size` cannot be safely cast to `usize`, throws `error.FileTooBig`
similarly to `File.readAllAlloc`.

While here, this commit also updates the docs of `Reader.readNoEof`
which were outdated, and swaps `inStream()` for  `reader()` in
`File.readAllAlloc` with the former being deprecated.