Add std.os.realpathat and std.fs.Dir.realpath
`std.os.realpathat` is similar to `std.os.realpath`, however, it accepts
a pair `fd_t, []const u8` of args and thus works out the realpath
of a relative path wrt to some opened file descriptor.

If the input pathname argument turns out to be an absolute path, this
function reverts to calling `realpath` on that pathname completely
ignoring the input file descriptor. This behaviour is standard in
POSIX and IMHO a good rule of thumb to follow.

If the input file descriptor was obtained using `std.fs.cwd()` call,
this function reverts to `std.os.getcwd()` to obtain the file
descriptor's path.

`std.fs.Dir.realpath` integrates `std.os.realpathat` with `std.fs.Dir`
but with dynamic memory allocator.