Restructured integer parsing to allow for most negative integer
Previously we simply negated the result of parsing the absolute value, which would overflow for the most negative integer. This version negates every digit individually, which fixes that issue at the cost of some performance. Perhaps there's a more efficient implementation, I'll be happy to implement it if anyone knows. Public interface is preserved, although `parseUnsignedError` has been renamed to `parseIntError`.